requirements.txt
Flask==2.3.3
TensorFlow==2.13.0
Keras==2.13.1
tensorflow-datasets==4.9.6
numpy<2
Pillow==10.0.0
Werkzeug==2.3.7
matplotlib==3.7.2
scikit-learn==1.3.0
